+++ tests/Classifiers/MultiLayerPerceptronTest.php
@@ -108,6 +108,22 @@ class MultiLayerPerceptronTest extends TestCase
         $this->assertGreaterThanOrEqual(0.5, $probabilities[1]['female']);
     }
 
+    public function test_partial_train()
+    {
+        $folds = $this->training->randomize()->stratifiedFold(3);
+
+        $this->estimator->train($folds[0]);
+
+        $this->estimator->partial($folds[1]);
+
+        $this->estimator->partial($folds[2]);
+
+        $predictions = $this->estimator->predict($this->testing);
+
+        $this->assertEquals($this->testing->label(0), $predictions[0]);
+        $this->assertEquals($this->testing->label(1), $predictions[1]);
+    }
+
     public function test_train_with_unlabeled()
     {
         $dataset = new Unlabeled([['bad']]);