+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 - Unreleased
     - Added Extra Tree classifier
-    - Committee Machine now supports influence
     - Random Forest now supports Extra Trees
     - New Decision Tree implementation
     - Added Canberra distance kernel
     - Committee Machine is now a Meta Estimator Ensemble
     - Added Bootstrap Aggregator Meta Estimator Ensemble
+    - Added Guassian Naive Bayes
+    - Naive Bayes classifiers are now Online Estimators
 
 - 0.0.2-alpha
     - Added Anomaly Detection